requirements.txt
streamlit
langchain
chromadb
openai
pdfplumber
yfinance
plotly
pandas
requests
python-dotenv
sentence-transformers
